refactor: icon button 으로 변경
