+++ src/ui/button/index.ts
@@ -1,3 +1,3 @@
 export { default as Button } from './base/Button';
 export { default as ImageUploadButton } from './upload/ImageUploadButton';
-export { default as ButtonCancel } from './button-cancel/ButtonCancel';
+export { default as IconButton } from './icon-button/IconButton';
