+++ src/ui/index.ts
@@ -4,5 +4,5 @@ export * from './carousel/index';
 export * from './chip/index';
 export * from './dialog/index';
 export * from './footer/index';
-export * from './product-card/index';
 export * from './navigation/index';
+export * from './product-card/index';
